Make a variable `unsigned int' that was previously `unsigned _long_ int', for whatever reason.
git-svn-id: https://svn.dealii.org/trunk@5131 0785d39b-7218-0410-832d-ea1e28bc413d

+++ deal.II/deal.II/source/grid/grid_out.cc
@@ -153,7 +153,7 @@ template <int dim>
 unsigned int GridOut::n_boundary_faces (const Triangulation<dim> &tria) const
 {
   typename Triangulation<dim>::active_face_iterator face, endf;
-  unsigned long int n_faces = 0;
+  unsigned int n_faces = 0;
 
   for (face=tria.begin_active_face(), endf=tria.end_face();
        face != endf; ++face)
